+++ changelog.md
@@ -2,9 +2,25 @@
 
 This is a log of all the changes made to StreamFusion
 
+## 1.0.2
+
+### Fixes and changes:
+* Fixed a library file read error on start-up
+* Updated homepage in package.json
+* "Manage playlist" -button now uses a different icon
+* Volume is now saved when it is changed. When you start StreamFusion, volume will be set to the value it was left at.
+
+### New features:
+* New scrollbar
+* A new menu system
+* New License page
+* New About page
+* New user options menu and system, that allows changing the default search platform selection, as well as color schemes
+* New color schemes, including increased contrast and user defined
+
 ## 1.0.1
 
-### Fixes:
+### Fixes and changes:
 * Misc. bugfixes
 
 ### New features: